src/application/use-cases/authenticate-user.ts
import { Injectable, UnauthorizedException } from '@nestjs/common'
import { compare } from 'bcryptjs'
import { UsersRepository } from '@/application/repositories/users-repository'

interface AuthenticateUserUseCaseRequest {
  email: string
  password: string
}

interface AuthenticateUserUseCaseResponse {
  userId: string
}

@Injectable()
export class AuthenticateUserUseCase {
  constructor(private usersRepository: UsersRepository) {}

  async execute({
    email,
    password,
  }: AuthenticateUserUseCaseRequest): Promise<AuthenticateUserUseCaseResponse> {
    const user = await this.usersRepository.findByEmail(email)

    if (!user) {
      throw new UnauthorizedException('Invalid credentials.')
    }

    const isPasswordValid = await compare(password, user.password)

    if (!isPasswordValid) {
      throw new UnauthorizedException('Invalid credentials.')
    }

    return { userId: user.id }
  }
}
